requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==11.1.0
django-filter==24.1
